requirements.txt
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
distlib==0.3.7
distro==1.8.0
dnspython==2.4.2
docker==6.1.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
exceptiongroup==1.1.3
fastapi==0.103.1
filelock==3.12.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
jsonschema==3.2.0
packaging==23.1
paho-mqtt==1.6.1
paramiko==3.3.1
platformdirs==3.10.0
pluggy==1.3.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pymongo==4.5.0
PyNaCl==1.5.0
pyrsistent==0.19.3
pytest==7.4.1
python-dotenv==0.21.1
PyYAML==5.4.1
redis==5.0.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
texttable==1.6.7
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
virtualenv==20.24.4
websocket-client==0.59.0
